refactor: split wizard steps into components
