+++ src/domain/splitter.ts
@@ -29,7 +29,7 @@ export type SplitFormValues = {
 };
 
 export const PARTICIPANT_NAME_MAX_LENGTH = 25;
-export const ITEM_NAME_MAX_LENGTH = 40;
+export const ITEM_NAME_MAX_LENGTH = 32;
 export const ITEM_AMOUNT_MAX_CENTS = 100_000_000;
 export const ITEM_AMOUNT_TOO_HIGH_MESSAGE =
   "Maximum is 1 000 000";
